Improve errors and documentation, use recv(2) to receive pings, ignore /inetutils-2.0
# Notes

Add notes for IP_RECVERR, IPPROTO_RAW, the fact that IP packets are recv(2)'ed with the header, the --verbose flag.

Add notes for ways to test this.

Add a reference to the Debian ping.

Add references about the checksum algorithm.

Add a to-do to include time in packets.

# Etc.

Make errors look more like the reference ping errors.

Receive replies.

# Minor changes

Set the sequence number, recompute the checksum in the loop, not outside of it.

Reduce the count of pings to one.